New method for discriminating between optimal implementations
This takes advantage of the change in behaviour of interworking branches
between ARMv6 and ARMv7 to distinguish which one we're running on. This makes
the test OS-independent (and in any case it's questionable how safe it is to
be making system calls when a function as fundamental as memcpy isn't
functional yet, so this is probably a safer approach).

The main downside here is that anyone trying the library on an ARMv7 platform
which lacks NEON (the usual example is nVidia Tegra 2) will find it doesn't
work. But then we're not pretending to offer an optimised implementation for
that anyway, and they might as well continue to use glibc.